Add LDAPS as protocol
Since the addition of IMAPS it is only another case to add for check_ssl_cert.
It has the advantage that it works with OpenSSL < 1.1.1 as LDAP requires an
option from OpenSSL that was only added in the 1.1.1 branch.

While LDAPS is considered deprecated in favour of StartTLS by at least
OpenLDAP and ApacheDS it is stil widely used, thus still useful.