Improve nmap speed to detect IPv6-only hosts
nmap performs a port scan to detect if it can resolve the address of the
host. If it fails it tries to force IPv6. This change tells nmap to scan
only the required port. This reduces the required time and reduces the
possibility to be falsely detected by an IDS.