Fix inconsistent output format in openssl x509 command (crl_uri)
The output of the `openssl x509` command with the `-text` and `-noout` options has changed between OpenSSL versions 1.1 and 3.0. Specifically, the output format of the "X509v3 CRL Distribution Points" section has changed. This commit updates the affected script to use a more robust and future-proof command that works on 1.1 and 3.0.