Add support to ignore unclean TLS shutdowns
The behaviour with respect to peers that fail to shutdown a TLS
connection cleanly has changed in OpenSSL 3. While the correct fix
for this is to update the application in question and make sure
they send the mandatory close_notify alert on shutdown, this is
not always feasible. Therefore add an option to ignore the
unexpected EOF that happens in this case.